refactor:테스트 기능별 분류
