--- src/test/java/racingcar/OutPutTest.java
@@ -1,4 +1,0 @@
-package racingcar;
-
-public class OutPutTest {
-}
